Issue #599 Moving DataSources fetching from query to services.
