+++ flexmeasures/api/v1/implementations.py
@@ -15,7 +15,7 @@ from flexmeasures.utils.entity_address_utils import (
 )
 from flexmeasures.data import db
 from flexmeasures.data.models.time_series import Sensor, TimedBelief
-from flexmeasures.data.queries.data_sources import get_or_create_source
+from flexmeasures.data.services.data_sources import get_or_create_source
 from flexmeasures.data.services.resources import get_sensors
 from flexmeasures.data.services.forecasting import create_forecasting_jobs
 from flexmeasures.api.common.responses import (
