+++ flexmeasures/cli/data_add.py
@@ -25,6 +25,7 @@ from flexmeasures.data.scripts.data_gen import (
     populate_initial_structure,
     add_default_asset_types,
 )
+from flexmeasures.data.services.data_sources import get_or_create_source
 from flexmeasures.data.services.forecasting import create_forecasting_jobs
 from flexmeasures.data.services.scheduling import make_schedule, create_scheduling_job
 from flexmeasures.data.services.users import create_user
@@ -53,8 +54,7 @@ from flexmeasures.data.schemas.generic_assets import (
 )
 from flexmeasures.data.models.generic_assets import GenericAsset, GenericAssetType
 from flexmeasures.data.models.user import User
-from flexmeasures.data.queries.data_sources import (
-    get_or_create_source,
+from flexmeasures.data.services.data_sources import (
     get_source_or_none,
 )
 from flexmeasures.utils import flexmeasures_inflection
