+++ flexmeasures/data/services/data_sources.py
@@ -1,12 +1,13 @@
-from __future__ import annotations
-
-from typing import Union, Optional
-
-from flask import current_app
 from flexmeasures import User
 from flexmeasures.data import db
 from flexmeasures.data.models.data_sources import DataSource
 from flexmeasures.data.models.user import is_user
+
+
+from flask import current_app
+
+
+from typing import Optional, Union
 
 
 def get_or_create_source(
